pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "arches_containers"
version = "0.1"
description = "A module for managing Arches container projects"
authors = [
    { name = "Andrew Jones", email = "andrew.jones@historicengland.org.uk" }
]
dependencies = [
    "python-slugify==7.0.0",
    "text-unidecode==1.3",
]

[project.scripts]
ac_create = "arches_containers.create_ac:main"
ac_manage = "arches_containers.manage_ac:main"